Complete Series (Video 7 to 10)
Again, forgot to do constant updates but this adds:
Updated Collision, Score UI, a Game Over State, Restarting Levels, a Level End Trigger, Advancing Levels, a Menu, and a Credit Screen that links back to the Menu. It only has one level for now, but soon I'll add the rest.